renamed Queries namespace to Query (better convention)
