+++ lib/evertils/common/query/backup.rb
@@ -8,27 +8,16 @@ require 'evertils/common/entity/sync'
 
 module Evertils
   module Common
-    module Queries
-      class Time
+    module Query
+      class Backup
 
-        def notes_created_in_range(start, finish)
+        #
+        # @since 0.2.8
+        def files(*files)
+          date = DateTime.now
+          note = Entity::Note.new
 
-        end
-
-        def notes_updated_in_range(start, finish)
-
-        end
-
-        def notebooks_created_in_range(start, finish)
-
-        end
-
-        def notebooks_updated_in_range(start, finish)
-
-        end
-
-        def last_sync
-
+          note.create("Backup: #{date.to_s}", '', 'Backup', files)
         end
 
       end
